+++ src/SeerConfigDialog.h
@@ -92,6 +92,15 @@ class SeerConfigDialog : public QDialog, protected Ui::SeerConfigDialogForm {
         void                                setAssemblySymbolDemagling                      (const QString& onoff);
         QString                             assemblySymbolDemagling                         () const;
 
+        void                                setAssemblyShowAddressColumn                    (bool flag);
+        bool                                assemblyShowAddressColumn                       () const;
+
+        void                                setAssemblyShowOffsetColumn                     (bool flag);
+        bool                                assemblyShowOffsetColumn                        () const;
+
+        void                                setAssemblyShowOpcodeColumn                     (bool flag);
+        bool                                assemblyShowOpcodeColumn                        () const;
+
         void                                setAssemblyRegisterFormat                       (const QString& format);
         QString                             assemblyRegisterFormat                          () const;